+++ Classes/BackEnd/AbstractEventMailForm.php
@@ -19,7 +19,7 @@ abstract class AbstractEventMailForm
     /**
      * @var string
      */
-    const MODULE_NAME = 'web_txseminarsM2';
+    const MODULE_NAME = 'web_seminars';
 
     /**
      * @var \Tx_Seminars_OldModel_Event the event which this e-mail form refers to